[pytree] moved 'pytree' parsing tools into its own subdirectory
