+++ yapf/yapflib/comment_splicer.py
@@ -25,7 +25,7 @@ from lib2to3 import pygram
 from lib2to3 import pytree
 from lib2to3.pgen2 import token
 
-from yapf.yapflib import pytree_utils
+from yapf.pytree import pytree_utils
 
 
 def SpliceComments(tree):
