+++ yapf/yapflib/format_token.py
@@ -21,8 +21,8 @@ import re
 
 from lib2to3.pgen2 import token
 
+from yapf.pytree import pytree_utils
 from yapf.yapflib import py3compat
-from yapf.yapflib import pytree_utils
 from yapf.yapflib import style
 from yapf.yapflib import subtypes
