+++ yapf/yapflib/yapf_api.py
@@ -36,6 +36,8 @@ import difflib
 import re
 import sys
 
+from yapf.pytree import pytree_unwrapper
+from yapf.pytree import pytree_utils
 from yapf.yapflib import blank_line_calculator
 from yapf.yapflib import comment_splicer
 from yapf.yapflib import continuation_splicer
@@ -43,8 +45,6 @@ from yapf.yapflib import errors
 from yapf.yapflib import file_resources
 from yapf.yapflib import identify_container
 from yapf.yapflib import py3compat
-from yapf.yapflib import pytree_unwrapper
-from yapf.yapflib import pytree_utils
 from yapf.yapflib import reformatter
 from yapf.yapflib import split_penalty
 from yapf.yapflib import style
