+++ yapftests/comment_splicer_test.py
@@ -16,9 +16,9 @@
 import textwrap
 import unittest
 
+from yapf.pytree import pytree_utils
 from yapf.yapflib import comment_splicer
 from yapf.yapflib import py3compat
-from yapf.yapflib import pytree_utils
 
 
 class CommentSplicerTest(unittest.TestCase):
