+++ yapftests/format_decision_state_test.py
@@ -16,9 +16,9 @@
 import textwrap
 import unittest
 
+from yapf.pytree import pytree_utils
 from yapf.yapflib import format_decision_state
 from yapf.yapflib import logical_line
-from yapf.yapflib import pytree_utils
 from yapf.yapflib import style
 
 from yapftests import yapf_test_helper
