+++ yapftests/pytree_unwrapper_test.py
@@ -16,7 +16,7 @@
 import textwrap
 import unittest
 
-from yapf.yapflib import pytree_utils
+from yapf.pytree import pytree_utils
 
 from yapftests import yapf_test_helper
