+++ yapftests/pytree_utils_test.py
@@ -19,7 +19,7 @@ from lib2to3 import pygram
 from lib2to3 import pytree
 from lib2to3.pgen2 import token
 
-from yapf.yapflib import pytree_utils
+from yapf.pytree import pytree_utils
 
 # More direct access to the symbol->number mapping living within the grammar
 # module.
