+++ yapftests/subtype_assigner_test.py
@@ -16,8 +16,8 @@
 import textwrap
 import unittest
 
+from yapf.pytree import pytree_utils
 from yapf.yapflib import format_token
-from yapf.yapflib import pytree_utils
 from yapf.yapflib import subtypes
 
 from yapftests import yapf_test_helper
