+++ yapftests/yapf_test_helper.py
@@ -17,14 +17,14 @@ import difflib
 import sys
 import unittest
 
+from yapf.pytree import pytree_unwrapper
+from yapf.pytree import pytree_utils
+from yapf.pytree import pytree_visitor
 from yapf.yapflib import blank_line_calculator
 from yapf.yapflib import comment_splicer
 from yapf.yapflib import continuation_splicer
 from yapf.yapflib import identify_container
 from yapf.yapflib import py3compat
-from yapf.yapflib import pytree_unwrapper
-from yapf.yapflib import pytree_utils
-from yapf.yapflib import pytree_visitor
 from yapf.yapflib import split_penalty
 from yapf.yapflib import style
 from yapf.yapflib import subtype_assigner
